1)Add GameScreen with functional from GamePad.class
2)GamePad.class is deleted
