+++ core/src/com/mygdx/game/MyGdxGame.java
@@ -1,10 +1,7 @@
 package com.mygdx.game;
 
-import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Game;
-import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.mygdx.game.gameScreens.GameScreen;
@@ -12,50 +9,21 @@ import com.mygdx.game.gameScreens.GameScreen;
 public class MyGdxGame extends Game {
     SpriteBatch batch;
     Texture img;
-	private Screen screen;
-	public static int GENERAL_WIDTH = 480; // 480
-	public static int GENERAL_HEIGHT = 720; // 720
-	public static float SCALE_FACTOR;
-	@Override
-	public void create () {
-		batch = new SpriteBatch();
-		img = new Texture("badlogic.jpg");
-		setScreen(new GameScreen());
-	}
+    public static int GENERAL_HEIGHT = 720; // 720
+    public static float SCALE_FACTOR;
 
-	@Override
-	public void render () {
-		Gdx.gl.glClearColor(1, 0, 0, 1);
-		Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-		batch.begin();
-		batch.draw(img, 0, 0);
-		batch.end();
-	}
-	
-	@Override
-	public void dispose () {
-		batch.dispose();
-		img.dispose();
-	}
+    @Override
+    public void create() {
+        batch = new SpriteBatch();
+        img = new Texture("badlogic.jpg");
+        setScreen(new GameScreen());
+    }
 
+    @Override
+    public void dispose() {
+        batch.dispose();
+        img.dispose();
+    }
 
-
-	/** Sets the current screen. {@link Screen#hide()} is called on any old screen, and {@link Screen#show()} is called on the new
-	 * screen, if any.
-	 * @param screen may be {@code null}
-	 */
-	public void setScreen (Screen screen) {
-		if (this.screen != null) this.screen.hide();
-		this.screen = screen;
-		if (this.screen != null) {
-			this.screen.show();
-			this.screen.resize(Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
-		}
-	}
-
-	/** @return the currently active {@link Screen}. */
-	public Screen getScreen () {
-		return screen;
-	}
 }
