+++ core/src/com/mygdx/game/gameScreens/MenuScreen.java
@@ -1,8 +1,6 @@
 package com.mygdx.game.gameScreens;
 
-import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.ScreenAdapter;
-import com.badlogic.gdx.assets.AssetManager;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.Stage;
@@ -15,7 +13,6 @@ import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.mygdx.game.MyGdxGame;
 import com.mygdx.game.assets.Assets;
-import com.mygdx.game.gameworld.gameobjects.touchpad.GamePad;
 
 import static com.mygdx.game.MyGdxGame.SCALE_FACTOR;
 
@@ -53,7 +50,7 @@ public class MenuScreen extends ScreenAdapter {
             @Override
             public void clicked(InputEvent event, float x, float y) {
                 batch.dispose();
-                game.setScreen(new GamePad());//TODO переделать GamePad в игровой экран
+                game.setScreen(new GameScreen());
             }
         });
     }
